backend/requirements.txt
fastapi
uvicorn
pydantic
requests
pyusb
pyo3
nuitka
zstandard
